pyproject.toml
[project]
name = "rtf-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
    "bcrypt>=5.0.0",
    "pre-commit>=4.3.0",
]
